// src/shared/enums/errorText.ts
export enum ErrorText {
  PARTNER_NAME_REQUIRED = 'Name is required',
  SALUTATION_REQUIRED = 'Salutation is required',
  FIRST_NAME_REQUIRED = 'First name is required',
  LAST_NAME_REQUIRED = 'Last name is required',
  JOB_TITLE_REQUIRED = 'Job Title is required',
  EMAIL_REQUIRED = 'Email is required',
  CORRECT_EMAIL = 'Email address is not correct',
  CORRECT_WEBSITE_URL = 'Website URL is not a valid fully-qualified http, https, or ftp URL',
  CORRECT_LOGO_URL = 'Logo URL is not a valid fully-qualified http, https, or ftp URL',
}